[FIX] removed second class from file
